102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,17 +7,17 @@
  */
 int main(void)
 {
-	long long int a, b, c, n;
+	long int a, b, c, n;
 
 	a = 1;
 	b = 2;
 	n = 3;
-	printf("%lld", a);
-	printf(", %lld", b);
+	printf("%lu", a);
+	printf(", %lu", b);
 	while (n <= 50)
 	{
 		c = a + b;
-		printf(", %lld", c);
+		printf(", %lu", c);
 		a = b;
 		b = c;
 		n++;
